+build is file level, also convert BenchmarkSize*
